pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatgpt-data-platform"
version = "0.1.0"
description = "End-to-end data platform simulating ChatGPT's data infrastructure"
requires-python = ">=3.10"
dependencies = [
    "pyspark>=3.5.0",
    "kafka-python-ng>=2.0.2",
    "faust-streaming>=0.11.0",
    "minio>=7.2.0",
    "pyarrow>=14.0.0",
    "snowflake-connector-python>=3.6.0",
    "snowflake-sqlalchemy>=1.5.0",
    "great-expectations>=0.18.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "faker>=22.0.0",
    "numpy>=1.26.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "pandas>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
generate-data = "src.generators.runner:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
